fix(core): widen the by-id API to Int64, and the guards that read a key back
Issue #333. The DataSet and REST by-id API took an Integer while everything
below it already took Int64 or TValue: TSessionAbstract<M>.Find has taken
Int64 since #324 and TDataSetBaseAdapter<M>.Find(Integer) did nothing but
`Result := FSession.Find(AID)`; OpenIDInternal and TSessionAbstract<M>.OpenID
both take a TValue and TContainerDataSet<M>.Open narrowed on the way in. The
32-bit limit lived only in the public entry points.

EIGHTEEN DECLARATIONS AND IMPLEMENTATIONS, RE-ENUMERATED ON 0546a51 rather
than taken from the issue, across Janus.Session.Abstract,
Janus.Container.DataSet(.Interfaces), Janus.DataSet.Base.Adapter,
Janus.Manager.DataSet, Janus.Session.RESTful, Janus.Server.RestObjectSet and
Janus.Server.RestObjectSet.Session. Every anchor was matched exactly once
before being rewritten.

WHY Int64 AND NOT AN OVERLOAD PAIR. Find(Integer) alongside Find(Int64) is
ambiguous at every untyped call site - `Find(10)` cannot pick one - so the
consumers that pass a literal would STOP compiling. Widening the single
parameter keeps all of them: an Integer argument promotes losslessly.
IntegerLiteral_StillCompilesAndIsUnchanged is the clause that holds that
down, and it passes a bare literal on purpose.

WHO BREAKS: NOBODY MEASURED. All seven test projects build with zero errors
and zero new warnings, which covers every descendant that overrides and every
implementer of IContainerDataSet inside this repository. THE PREMISE THAT
THIRD PARTIES IMPLEMENT THESE INTERFACES IS NOT MEASURABLE FROM HERE AND IS
DECLARED NOT MEASURED.

THREE NARROWINGS THAT SAT BEHIND THE API, and two of them behind a guard that
contradicted itself: cINTEGERKINDS admits ftLargeint, so
_AutoIncKeyIsGenerated and _RowKeyIsUngenerated ACCEPTED a 64-bit key and then
compared `LField.AsInteger = cAutoIncNotGenerated` (-1). TLargeintField.
GetAsInteger returns `Integer(L)`, a hard truncation nothing checks - measured
from the RTL source under #324 - so the ordinary key 4294967295 came back as
exactly -1 and a key that EXISTS was declared ungenerated and silently not
cascaded. Both now read AsLargeInt. The third,
TSessionRestFul<M>.Delete(const AObject: M), took the first key column with
.AsInteger and now takes .AsInt64.

RED-FIRST, AND THE FIRST VERSION OF THE PREDICTION WAS WRONG. It said the
Integer entry point would truncate silently; the run said Range check error,
because DCC_RangeChecking is true in these Debug builds. All three wide
clauses ERRORED on the unwidened tree and the Integer control was green. No
truncated literal was ever observed at that layer and none is claimed. The
guard's truncation IS silent, because that one is an RTL cast and not a
compiler-inserted conversion - both are recorded.

MUTATIONS, all with the altered line echoed by dcc32 as W1054 in the same
build: Find truncated to 32 bits -> 2 red; Open truncated -> 1 red;
_AutoIncKeyIsGenerated reverted to AsInteger -> 0 red, A SURVIVOR, which is
why WideAutoIncKey_ThatTruncatesToThePlaceholder_IsStillPropagated and its
cascade model were written; re-run against it -> 1 red.

ONE SURVIVOR REMAINS AND IS DECLARED, NOT HIDDEN: the same revert on
TRESTDataSetAdapter<M>._RowKeyIsUngenerated leaves RESTfulDriver at 134/0/0.
It is reachable only from a REST adapter over a model with an autoinc
ftLargeint PRIMARY KEY, and such a fixture must be registered in
Janus.Tests.RESTfulDriver.dpr, which the #323 frontier is editing. Left alone
deliberately rather than raced. Recorded over both guards.

Measured on this tree: Units 628/0/0, LiveBindings 31, RESTfulDriver 134,
RESTHorse 149, RESTMARS 33, RESTWiRL 19, RESTOracle 12 errored (basal, OCI
absent).

+++ Source/Core/Janus.Session.Abstract.pas
@@ -73,7 +73,7 @@ type
     procedure Update(const AObject: M; const AKey: String); overload; virtual;
     procedure Update(const AObjectList: TObjectList<M>); overload; virtual; abstract;
     procedure Delete(const AObject: M); overload; virtual;
-    procedure Delete(const AID: Integer); overload; virtual; abstract;
+    procedure Delete(const AID: Int64); overload; virtual; abstract;
     procedure LoadLazy(const AOwner, AObject: TObject); virtual;
     procedure InjectLazyProxies(const AObject: TObject); virtual;
     procedure NextPacketList(const AObjectList: TObjectList<M>); overload; virtual;
